Crate yarem

Yet Another Rust Error Macro

A simple but relatively easy to use Rust error macro. New errors can be defined using the error! macro as follows

#[macro_use] extern crate yarem;
/// A comment for this error that will be visible in the docs
error!(ErrorName, "The message that will be shown to the user");

In addition, errors can be combined using the errors! macro

#[macro_use] extern crate yarem;
// Names should be imported with a unique name
use std::io::{Error as IOError};
 
error!(CustomError, "A custom error");

errors!(IoOrCustom; IOError, CustomError);

This can then be used in the Result objects of functions, for example

#[macro_use] extern crate yarem;
use std::io::{Error as IOError};
use std::fs::File;
 
error!(CustomError, "A custom error");
errors!(IoOrCustom; IOError, CustomError);

fn bar() -> Result<(), CustomError> { Ok(()) }

fn foo() -> Result<(), IoOrCustom> {
  File::open("foo/foo")?;
  bar()?;
  Ok(())
}

Macros

error
Define a new error with an appropriate error message as its description, e.g., error!(A, "message") expands to

errors
Define a new error that is an enumeration over all the possible values, i.e, works as an algebraic union of the types. Error, Display and From are implemented for this new error type. For example errors!(X; A, B) generates: